requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-extensions==2.1.5
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.1
djangorestframework-filters==0.10.2.post0
djangorestframework-simplejwt==4.0.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==5.4.1
pluggy==0.8.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.1.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
zope.interface==4.6.0
gunicorn==19.9.0
django-silk==3.0.2
django-prometheus==1.0.15
